requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-dateutil>=2.8.0
tqdm>=4.66.0
Pillow>=10.0.0
whoosh>=2.7.0
click>=8.1.0
tenacity>=8.2.0
html2text>=2020.1.16
flask>=2.3.0
flask-cors>=4.0.0